+++ .gitignore
@@ -36,3 +36,5 @@ yarn-error.log*
 
 # webstorm
 .idea
+
+.vscode
